% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setCache.R
\name{cache-cache}
\alias{cache-cache}
\alias{setCache}
\alias{getCache}
\title{Helper functions for package caching}
\usage{
setCache(directory = rappdirs::user_cache_dir(getOption("pkgName")),
  verbose = TRUE, ask = interactive())

getCache()
}
\arguments{
\item{directory}{location on disk intended for caching files}

\item{verbose}{logical (default FALSE) whether to report procedural steps}

\item{ask}{logical (default `interactive()``) whether to ask the user to
confirm the cache location}
}
\value{
setCache: Modifies 'pkgName' and <pkg>_cache options and creates
  directory if not available

A BiocFileCache instance for the package specified in the 'pkgName'
  option
}
\description{
Both setCache and getCache allow maintainers / developers to
specify cache locations for the package they wish to implement a caching
mechanism. This avoids the maintainer the use of BiocFileCache directly by
providing a simple option called `pkgName` which indicates the name of the
package being developed.
}
\examples{
## options(pkgName = "cacheur")
setCache()


}
