% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadHDF5FromS3.R
\name{loadDelayedSEFromS3}
\alias{loadDelayedSEFromS3}
\title{Pull HDF5Array files from the associated S3 bucket}
\usage{
loadDelayedSEFromS3(bucket = "multiassayexperiments",
  dataname = "example", verbose = FALSE, force = FALSE)
}
\arguments{
\item{bucket}{A string indicating the S3 bucket name}

\item{dataname}{The name of the folder as part of the bucket path}

\item{verbose}{logical (default FALSE) whether to report procedural steps
during download and load}

\item{force}{logical (default FALSE) whether to re-download and force load
resources from S3}
}
\description{
This function downloads and caches data from an Amazon S3 bucket and
loads it as a SummarizedExperiment with an HDF5Array
}
